% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_plans.R
\name{download_plans}
\alias{download_plans}
\title{Download conservation plans using URL}
\usage{
download_plans(
  dir = "./data-raw/imports",
  country = NULL,
  states = NULL,
  inverse.states = NULL
)
}
\arguments{
\item{dir}{Directory within which the downloaded reports will be saved. Defaults to "./data-raw/imports/". If it does not already exist, will create this subdirectory.}

\item{country}{Defaults to all regions. Country abbreviation, one or more of c("US", "CAN"), by which the plans will be downloaded.}

\item{states}{Defaults to all states. State or province abbreviation, one of c("US", "CAN"), by which the plans will be downloaded.}

\item{inverse.states}{Defaults to NULL. If populated and argument 'state' is not defined, will download everything EXCEPT those states/regions provided here.}
}
\description{
A helper function for downloading conservation plans and reports of all or a select number (see paramter: 'plans') appearing in data, 'plan_urls'.
}
