% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_text_from_plans.R
\name{extract_text_from_plans}
\alias{extract_text_from_plans}
\title{extract_text_from_plans
Function to bulk-extract and import text from plans and reports}
\usage{
extract_text_from_plans(dir = "./data-raw/imports", all = TRUE)
}
\description{
extract_text_from_plans
Function to bulk-extract and import text from plans and reports
}
