% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plan_urls}
\alias{plan_urls}
\title{plan_urls lookup table}
\format{
An object of class \code{data.frame} with 4 rows and 9 columns.
}
\usage{
plan_urls
}
\description{
Links and basic metadata for conservation plans, reports, including links to downloadable files where applicable.
}
\keyword{datasets}
